+++ app/controllers/users_controller.rb
@@ -52,6 +52,8 @@ class UsersController < ApplicationController
 
   def index
     @users = User.ordered.paginate(page: params[:page])
+    @users = @users.search(params[:q])
+    @users = @users.by_agent(params[:agent] == '1') unless params[:agent].blank?
   end
 
   def new