+++ db/schema.rb
@@ -11,15 +11,15 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150807135450) do
+ActiveRecord::Schema.define(version: 20150821125309) do
 
   create_table "attachments", force: :cascade do |t|
     t.integer  "attachable_id"
-    t.string   "attachable_type",   limit: 255
+    t.string   "attachable_type"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "file_file_name",    limit: 255
-    t.string   "file_content_type", limit: 255
+    t.string   "file_file_name"
+    t.string   "file_content_type"
     t.integer  "file_file_size"
     t.datetime "file_updated_at"
   end
@@ -46,32 +46,32 @@ ActiveRecord::Schema.define(version: 20150807135450) do
   create_table "labelings", force: :cascade do |t|
     t.integer  "label_id"
     t.integer  "labelable_id"
-    t.string   "labelable_type", limit: 255
+    t.string   "labelable_type"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
   add_index "labelings", ["label_id", "labelable_id", "labelable_type"], name: "unique_labeling_label", unique: true
   add_index "labelings", ["label_id"], name: "index_labelings_on_label_id"
-  add_index "labelings", ["labelable_id", "labelable_type"], name: "index_labelings_on_labelable_id_and_labelable_type"
+  add_index "labelings", ["labelable_type", "labelable_id"], name: "index_labelings_on_labelable_type_and_labelable_id"
 
   create_table "labels", force: :cascade do |t|
-    t.string   "name",       limit: 255
+    t.string   "name"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "color",      limit: 255
+    t.string   "color"
   end
 
   create_table "notifications", force: :cascade do |t|
     t.integer  "notifiable_id"
-    t.string   "notifiable_type", limit: 255
+    t.string   "notifiable_type"
     t.integer  "user_id"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
   add_index "notifications", ["notifiable_id", "notifiable_type", "user_id"], name: "unique_notification", unique: true
-  add_index "notifications", ["notifiable_id", "notifiable_type"], name: "index_notifications_on_notifiable_id_and_notifiable_type"
+  add_index "notifications", ["notifiable_type", "notifiable_id"], name: "index_notifications_on_notifiable_type_and_notifiable_id"
   add_index "notifications", ["user_id"], name: "index_notifications_on_user_id"
 
   create_table "replies", force: :cascade do |t|
@@ -80,8 +80,8 @@ ActiveRecord::Schema.define(version: 20150807135450) do
     t.datetime "updated_at"
     t.integer  "ticket_id"
     t.integer  "user_id"
-    t.string   "message_id",   limit: 255
-    t.string   "content_type", limit: 255,        default: "html"
+    t.string   "message_id"
+    t.string   "content_type",                    default: "html"
   end
 
   add_index "replies", ["message_id"], name: "index_replies_on_message_id"
@@ -89,11 +89,11 @@ ActiveRecord::Schema.define(version: 20150807135450) do
   add_index "replies", ["user_id"], name: "index_replies_on_user_id"
 
   create_table "rules", force: :cascade do |t|
-    t.string   "filter_field",     limit: 255
-    t.integer  "filter_operation",             default: 0, null: false
-    t.string   "filter_value",     limit: 255
-    t.integer  "action_operation",             default: 0, null: false
-    t.string   "action_value",     limit: 255
+    t.string   "filter_field"
+    t.integer  "filter_operation", default: 0, null: false
+    t.string   "filter_value"
+    t.integer  "action_operation", default: 0, null: false
+    t.string   "action_value"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
@@ -115,14 +115,14 @@ ActiveRecord::Schema.define(version: 20150807135450) do
   end
 
   create_table "tickets", force: :cascade do |t|
-    t.string   "subject",             limit: 255
+    t.string   "subject"
     t.text     "content",             limit: 1073741823
     t.datetime "created_at"
     t.datetime "updated_at"
     t.integer  "assignee_id"
-    t.string   "message_id",          limit: 255
+    t.string   "message_id"
     t.integer  "user_id"
-    t.string   "content_type",        limit: 255,        default: "html"
+    t.string   "content_type",                           default: "html"
     t.integer  "status",                                 default: 0,      null: false
     t.integer  "priority",                               default: 0,      null: false
     t.integer  "to_email_address_id"
@@ -141,23 +141,23 @@ ActiveRecord::Schema.define(version: 20150807135450) do
   create_table "users", force: :cascade do |t|
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "email",                  limit: 255, default: "",          null: false
-    t.string   "encrypted_password",     limit: 255, default: "",          null: false
-    t.string   "reset_password_token",   limit: 255
+    t.string   "email",                  default: "",          null: false
+    t.string   "encrypted_password",     default: "",          null: false
+    t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",                      default: 0
+    t.integer  "sign_in_count",          default: 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip",     limit: 255
-    t.string   "last_sign_in_ip",        limit: 255
-    t.boolean  "agent"
+    t.string   "current_sign_in_ip"
+    t.string   "last_sign_in_ip"
+    t.boolean  "agent",                  default: false,       null: false
     t.text     "signature"
-    t.boolean  "notify",                             default: true
-    t.string   "authentication_token",   limit: 255
-    t.string   "time_zone",              limit: 255, default: "Amsterdam"
-    t.integer  "per_page",                           default: 30,          null: false
-    t.string   "locale",                 limit: 255
+    t.boolean  "notify",                 default: true
+    t.string   "authentication_token"
+    t.string   "time_zone",              default: "Amsterdam"
+    t.integer  "per_page",               default: 30,          null: false
+    t.string   "locale"
   end
 
   add_index "users", ["email"], name: "index_users_on_email", unique: true